+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Add ability to review and revoke particular logged in user sessions
+- Add ability to change password from user settings screen
 
 ### Removed